backend/requirements.txt
# Core Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.32


# ML & Stats
scikit-learn>=1.3.0
joblib>=1.2.0
vaderSentiment>=3.3.2

# RAG & AI Models
sentence-transformers>=2.2.2
openai>=1.10.0
anthropic>=0.18.0
google-genai>=0.6.0
ollama>=0.1.0

# News & Web Scraping
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API Clients
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0
httpx>=0.25.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Visualization


# Logging
loguru>=0.7.2
prometheus-client>=0.19.0
